--- cypress/support/fragments/data-export/settingsDataExport.js
@@ -1,9 +1,0 @@
-export default {
-  goToMappingProfiles() {
-    cy.visit('/settings/data-export/mapping-profiles');
-  },
-
-  goToJobProfiles() {
-    cy.visit('/settings/data-export/job-profiles');
-  },
-};
